Code quality/style improvements
Removed the need for individual SQL queries to create tables and
insert items. For this purpose, Table and Tuple extensions are added,
which provide the possibility to retrieve SQL queries. This adds a
bit of overhead in the number of classes that were needed, but
improves the code quality and makes future changes to tuples or
tables easier. Also added javadoc comments.