Splash screen working + multithreading
splash screen seems to be working now. The heavy tasks must be done on a
seperate thread or massive problems will occur (GUI not responding etc).
The progress bar and the text under it updates now correctly on the GUI
while keeping the GUI responsive with the code underlying it running on a
seperate thread.