Fix not appearing of scrollbar on zoom
The vertical scrollbar did not appear when zooming in, which caused
a large part of the graph to be invisible to the user. This is now
fixed by wrapping the graph in an outer Group, which is necessary
when the scrolling needs to be based on the visual bounds of the
node (according to the JavaDoc for the ScrollPane class).