Modified algorithm
The previously used algorithm to find bubbles turned out to be hard
to modify so that the level of a bubble can be found. This new
algorithm only locates the innermost bubbles (bubbles that do not
contain inner bubbles). Some bubbles might not be detected yet,
but this will be modified later.